:sparkles: Move Migration Waves behind a disabled feature flag + some renaming (#790)
Mostly this PR makes it possible for us to merge
`feature-migration-waves` to main in a non-disruptive way by disabling
the waves route and nav item with a boolean feature flag. I also renamed
some things and replaced my TODO comments about introducing a
`MigrationWave` type with the existing `Wave` type I had forgotten we
already had.

---------
