+++ client/src/app/FeatureFlags.ts
@@ -1,0 +1,3 @@
+export const FEATURES_ENABLED = {
+  migrationWaves: false,
+};
