--- client/src/app/shared/components/composable-app-table/index.ts
@@ -1,1 +1,0 @@
-export { ComposableAppTable } from "./composable-app-table";
